+++ .storybook/main.js
@@ -1,6 +1,6 @@
 module.exports = {
     "stories": [
-        "../**/*.stories.@(js|jsx|ts|tsx)"
+        "../src/**/*.stories.@(js|jsx|ts|tsx)"
     ],
     "addons": [
         {
